:recycle: refactor ptt crawler feature
